reorg: move bbox and coder
